authOptins 분리 및 적용
