+++ src/app/api/auth/[...nextauth]/auth.ts
@@ -1,0 +1,50 @@
+import { NextAuthOptions } from "next-auth";
+import { PrismaAdapter } from "@next-auth/prisma-adapter";
+import GoogleProvider from "next-auth/providers/google";
+import NaverProvider from "next-auth/providers/naver";
+import KakaoProvider from "next-auth/providers/kakao";
+import prisma from "@/db";
+
+const ONE_HOUR_IN_SECONDS = 3600;
+const ONE_DAY_IN_SECONDS = 86400;
+
+export const authOptions: NextAuthOptions = {
+  session: {
+    strategy: "jwt" as const,
+    maxAge: ONE_DAY_IN_SECONDS,
+    updateAge: ONE_HOUR_IN_SECONDS,
+  },
+  adapter: PrismaAdapter(prisma),
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID || "",
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET || "",
+    }),
+    NaverProvider({
+      clientId: process.env.NAVER_CLIENT_ID || "",
+      clientSecret: process.env.NAVER_CLIENT_SECRET || "",
+    }),
+    KakaoProvider({
+      clientId: process.env.KAKAO_CLIENT_ID || "",
+      clientSecret: process.env.KAKAO_CLIENT_SECRET || "",
+    }),
+  ],
+  pages: {
+    signIn: "/users/login",
+  },
+  callbacks: {
+    session: ({ session, token }) => ({
+      ...session,
+      user: {
+        ...session.user,
+        id: token.sub,
+      },
+    }),
+    jwt: async ({ user, token }) => {
+      if (user) {
+        token.sub = user.id;
+      }
+      return token;
+    },
+  },
+};
